+++ examples/hooks/package.json
@@ -19,6 +19,7 @@
   "devDependencies": {
     "@types/node": "^17.0.15",
     "node-fetch": "^2.6.7",
+    "@types/node-fetch": "^2.6.2",
     "npm-run-all": "^4.1.5",
     "open-cli": "^7.0.1",
     "ts-node": "^10.8.0",
