+++ examples/nextjs-swr/package.json
@@ -26,6 +26,7 @@
   "devDependencies": {
     "@types/node": "^17.0.15",
     "@types/react": "^18.0.6",
+    "@types/node-fetch": "^2.6.2",
     "node-fetch": "^2.6.7",
     "npm-run-all": "^4.1.5",
     "open-cli": "^7.0.1",
