+++ examples/nextjs/package.json
@@ -24,6 +24,7 @@
   },
   "devDependencies": {
     "@types/node": "^17.0.15",
+    "@types/node-fetch": "^2.6.2",
     "@types/react": "^18.0.6",
     "node-fetch": "^2.6.7",
     "npm-run-all": "^4.1.5",
