+++ examples/openapi-mocking/jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+	roots: ['<rootDir>/test/'],
+	transform: {
+		'^.+\\.tsx?$': 'ts-jest',
+	},
+	testRegex: '(/__tests__/.*|\\.(test|spec))\\.[tj]sx?$',
+	moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
+};
