+++ examples/openapi-mocking/package.json
@@ -4,9 +4,9 @@
   "description": "",
   "scripts": {
     "start": "wunderctl up --debug",
-    "generate": "wunderctl generate up --debug",
+    "generate": "wunderctl generate --debug",
     "check": "tsc --noEmit",
-    "test": "ava"
+    "test": "jest"
   },
   "author": "",
   "license": "ISC",
@@ -15,8 +15,13 @@
     "graphql": "^16.3.0"
   },
   "devDependencies": {
-    "ava": "^5.1.0",
-    "ts-node": "^10.9.1",
+    "@jest/globals": "^29.3.1",
+    "@types/node": "^17.0.15",
+    "@types/node-fetch": "^2.6.2",
+    "jest": "^29.3.1",
+    "node-fetch": "^2.6.7",
+    "ts-jest": "^29.0.3",
+    "ts-node": "^10.8.0",
     "typescript": "^4.8.2"
   },
   "ava": {
