+++ examples/simple/package.json
@@ -18,6 +18,7 @@
     "@jest/globals": "^29.3.1",
     "@types/node": "^14.14.37",
     "@types/node-fetch": "^2.6.2",
+    "node-fetch": "^2.6.7",
     "jest": "^29.3.1",
     "ts-jest": "^29.0.3",
     "typescript": "^4.8.2"
